Move prepending to main sidebar component
When the prepending functionality was in a component within the
sidebar, the state didn't get updated correctly if the sidebar was
closed. This way even if the Post sidebar is open and you're changing
e.g. a post's categories, log notes will still get loaded into state.
